HW-5 - Refactoring Project
