+++ class13/Sangmeebee/MyProject/app/src/main/AndroidManifest.xml
@@ -5,14 +5,14 @@
     <uses-permission android:name="android.permission.INTERNET"/>
 
     <application
-        android:name=".data.App"
+        android:name=".data.sharedPreferences.App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.main.MainActivity">
+        <activity android:name=".ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
