+++ class13/Sangmeebee/MyProject/app/src/main/java/com/example/myproject/data/sharedPreferences/Preferences.kt
@@ -1,4 +1,4 @@
-package com.example.myproject.data
+package com.example.myproject.data.sharedPreferences
 
 import android.app.Activity
 import android.content.Context
