+++ class13/Sangmeebee/MyProject/app/src/main/res/values/strings.xml
@@ -1,3 +1,7 @@
 <resources>
     <string name="app_name">MyProject</string>
+    <string name="call_success">성공</string>
+    <string name="call_empty">제목을 입력하세요.</string>
+    <string name="call_error">네트워크 에러발생</string>
+    <string name="call_result_empty">결과가 없습니다.</string>
 </resources>
